+++ odoo_health_check/models/__init__.py
@@ -3,3 +3,4 @@ from . import ir_cron
 from . import health_check_result
 from . import health_check_dashboard
 from . import res_config_settings
+from . import mail_render_mixin